refactor(entity): Move sea_skipper impls to extend
Renames the private `behavior` module to `extend`.

Moves some manual trait implementations from `sea_skipper` into
`extend` to reduce toil when updating generated code.
