--- entity/src/behavior/entry.rs
@@ -1,15 +1,0 @@
-//! Logic for [`entry`][crate::entry] entity that is not generated by [`sea_orm`].
-
-use sea_orm::{prelude::*, Set};
-
-use crate::{entry, Uuid};
-
-impl ActiveModelBehavior for entry::ActiveModel {
-    /// Create a new `entry::ActiveModel` with a random id and default values.
-    fn new() -> Self {
-        Self {
-            id: Set(Uuid::new_v4()),
-            ..ActiveModelTrait::default()
-        }
-    }
-}
