+++ entity/src/drug.rs
@@ -3,7 +3,7 @@
 use crate::Uuid;
 
 use sea_orm::entity::prelude::*;
-use sea_skipper::{DeriveNewModel, Location, Resource};
+use sea_skipper::DeriveNewModel;
 use serde::{Deserialize, Serialize};
 
 #[derive(
@@ -26,23 +26,5 @@ pub enum Relation {
 impl Related<super::entry::Entity> for Entity {
     fn to() -> RelationDef {
         Relation::Entry.def()
-    }
-}
-
-// TODO: It's not ideal to implement these manually here because this is generated code.
-// This could be derived.
-impl Resource for Entity {
-    type ActiveModel = ActiveModel;
-    type Data = Model;
-    type Id = Uuid;
-}
-
-// TODO: It's not ideal to implement these manually here because this is generated code.
-// This could be derived here, or it may make more sense for `entity` to become a sub-module of
-// `edihkal` since that's where the route is defined. (But that has other considerations for model
-// definitions / dependencies for other crates e.g. `edihkal-client`.)
-impl Location for Model {
-    fn location(&self) -> String {
-        format!("/drugs/{}", self.id)
     }
 }
