+++ entity/src/entry.rs
@@ -1,7 +1,7 @@
 //! `SeaORM` Entity. Generated by sea-orm-codegen 0.11.2
 
 use sea_orm::entity::prelude::*;
-use sea_skipper::{DeriveNewModel, Location, Resource};
+use sea_skipper::DeriveNewModel;
 use serde::{Deserialize, Serialize};
 
 use crate::Uuid;
@@ -33,23 +33,5 @@ pub enum Relation {
 impl Related<super::drug::Entity> for Entity {
     fn to() -> RelationDef {
         Relation::Drug.def()
-    }
-}
-
-// TODO: It's not ideal to implement these manually here because this is generated code.
-// This could be derived.
-impl Resource for Entity {
-    type ActiveModel = ActiveModel;
-    type Data = Model;
-    type Id = Uuid;
-}
-
-// TODO: It's not ideal to implement these manually here because this is generated code.
-// This could be derived here, or it may make more sense for `entity` to become a sub-module of
-// `edihkal` since that's where the route is defined. (But that has other considerations for model
-// definitions / dependencies for other crates e.g. `edihkal-client`.)
-impl Location for Model {
-    fn location(&self) -> String {
-        format!("/entries/{}", self.id)
     }
 }
