+++ entity/src/extend/entry.rs
@@ -1,0 +1,32 @@
+//! Logic for [`entry`][crate::entry] entity that is not generated by [`sea_orm`].
+
+use sea_orm::{prelude::*, Set};
+use sea_skipper::{Location, Resource};
+
+use crate::{entry::*, Uuid};
+
+impl ActiveModelBehavior for ActiveModel {
+    /// Create a new `entry::ActiveModel` with a random id and default values.
+    fn new() -> Self {
+        Self {
+            id: Set(Uuid::new_v4()),
+            ..ActiveModelTrait::default()
+        }
+    }
+}
+
+// TODO: This could easily be derived.
+impl Resource for Entity {
+    type ActiveModel = ActiveModel;
+    type Data = Model;
+    type Id = Uuid;
+}
+
+// XXX: This may make more sense for `entity` to become a sub-module of
+// `edihkal` since that's where the route is defined. (But that has other considerations for model
+// definitions / dependencies for other crates e.g. `edihkal-client`.
+impl Location for Model {
+    fn location(&self) -> String {
+        format!("/entries/{}", self.id)
+    }
+}
