+++ entity/src/lib.rs
@@ -110,7 +110,7 @@ pub mod drug;
 pub mod entry;
 
 // The code below is not generated by sea-orm-codegen and may need to be restored after updating the generated entities.
-mod behavior;
+mod extend;
 pub mod uuid;
 
 pub use crate::uuid::UuidV4 as Uuid;
